Refactors mock data to remove cycle, Adds pass through to chat model
